* GWorkspace/Preferences/PrefController.h: specify type of popUp outlet to fix this dangerous warning: Preferences/PrefController.m:82:5: warning: multiple methods named ‘-removeItemAtIndex:’ found [enabled by default] /usr/local/include/AppKit/NSMenu.h:569:1: note: using ‘-(void)removeItemAtIndex:(int)index’ /usr/local/include/AppKit/NSToolbar.h:102:1: note: also found ‘-(void)removeItemAtIndex:(NSInteger)index’ * GWorkspace/Desktop/GWDesktopView.m: * Inspector/ContentViewers/NSColorViewer/NSColorViewer.m: use CGFloat instead of float for -[NSColor getRed:...]
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gworkspace/trunk@34122 72102866-910b-0410-8b05-ffd578937521

+++ Inspector/ContentViewers/NSColorViewer/NSColorViewer.m
@@ -145,8 +145,8 @@
 
   if (c && [c isKindOfClass: [NSColor class]]) {
     NSColor *color = [c colorUsingColorSpaceName: NSDeviceRGBColorSpace];
-    float red = 0.0, green = 0.0, blue = 0.0, alpha = 0.0;
-    float hue = 0.0, saturation = 0.0, brightness = 0.0;
+    CGFloat red = 0.0, green = 0.0, blue = 0.0, alpha = 0.0;
+    CGFloat hue = 0.0, saturation = 0.0, brightness = 0.0;
     
     if (valid == NO) {
       valid = YES;
